cleanup: fixed numerous compiler warnings
 - Fixed some argument warning/error issues
 - Fixed int-to-pointer and pointer-to-int warnings
 - Suppress few benign warnings from KVS SDK sources
 - Fixed redefined macros, unused vars issues
 - Upgraded esp_video version to fix dependency errors on latest IDF
 - Removed json dependency from app_webrtc
 - media_stream: Fixed deprecated i2s driver for esp32s3
 - media_stream: Fixed deprecated variable names in app_camera

+++ esp_port/components/kvs_utils/include/platform_esp32.h
@@ -88,7 +88,7 @@ typedef UINT64 TID;
 typedef TID* PTID;
 
 #ifndef INVALID_TID_VALUE
-#define INVALID_TID_VALUE ((UINT64) NULL)
+#define INVALID_TID_VALUE ((UINT64) (uintptr_t) NULL)
 #endif
 
 #ifndef IS_VALID_TID_VALUE
@@ -99,7 +99,7 @@ typedef TID* PTID;
 typedef UINT64 MUTEX;
 
 #ifndef INVALID_MUTEX_VALUE
-#define INVALID_MUTEX_VALUE ((UINT64) NULL)
+#define INVALID_MUTEX_VALUE ((UINT64) (uintptr_t) NULL)
 #endif
 
 #ifndef IS_VALID_MUTEX_VALUE
